update routes middleware

+++ routes/web.php
@@ -38,11 +38,11 @@ Route::controller(PostController::class)->group(function () {
     Route::get('/posts/{slug}', 'show')->name('posts.show')->middleware(['EnsureAnonymousSessionTokenExists']);
 
     // Admin Post Routes
-    Route::get('/admin/posts', 'adminIndex')->name('admin.posts.index');
+    Route::get('/admin/posts', 'adminIndex')->name('admin.posts.index')->middleware(['auth']);
     Route::get('/admin/posts/create', 'create')->name('admin.posts.create')->middleware(['auth']);
     Route::post('/admin/posts/create', 'store')->name('admin.posts.store')->middleware(['auth']);
-    Route::get('/admin/posts/{slug}/edit', 'edit')->name('admin.posts.edit');
-    Route::post('/admin/posts/{slug}/edit', 'update')->name('admin.posts.update');
+    Route::get('/admin/posts/{slug}/edit', 'edit')->name('admin.posts.edit')->middleware(['auth']);
+    Route::post('/admin/posts/{slug}/edit', 'update')->name('admin.posts.update')->middleware(['auth']);
     Route::post('/admin/posts/{id}/private', 'makePrivate')->name('admin.posts.private')->middleware(['auth']);
     Route::post('/admin/posts/{id}/publish', 'makePublic')->name('admin.posts.publish')->middleware(['auth']);
     Route::post('/admin/posts/create/banner-upload', 'uploadBannerImage')->name('admin.posts.upload-banner-image')->middleware(['auth']);
@@ -58,10 +58,10 @@ Route::controller(CategoryController::class)->group(function () {
     Route::get('categories/{slug}', 'show')->name('categories.show');
 
     // Admin Category Routes
-    Route::get('/admin/categories', 'adminIndex')->name('admin.categories.index');
-    Route::get('/admin/categories/create', 'create')->name('admin.categories.create');
-    Route::post('/admin/categories/create', 'store')->name('admin.categories.store');
-    Route::delete('/admin/categories/{id}/delete', 'destroy')->name('admin.categories.destroy');
+    Route::get('/admin/categories', 'adminIndex')->name('admin.categories.index')->middleware(['auth']);
+    Route::get('/admin/categories/create', 'create')->name('admin.categories.create')->middleware(['auth']);
+    Route::post('/admin/categories/create', 'store')->name('admin.categories.store')->middleware(['auth']);
+    Route::delete('/admin/categories/{id}/delete', 'destroy')->name('admin.categories.destroy')->middleware(['auth']);
 
 });
 
@@ -73,10 +73,10 @@ Route::controller(ImageController::class)->group(function () {
     Route::get('/gallery', 'gallery')->name('gallery.index');
 
     // Admin Image Routes
-    Route::get('/admin/images', 'index')->name('admin.images.index');
-    Route::get('/images/create', 'create')->name('images.create');
-    Route::post('/images/create', 'store')->name('images.store');
-    Route::delete('/images/{id}/delete', 'destroy')->name('images.destroy');
+    Route::get('/admin/images', 'index')->name('admin.images.index')->middleware(['auth']);
+    Route::get('/admin/images/create', 'create')->name('images.create')->middleware(['auth']);
+    Route::post('/admin/images/create', 'store')->name('images.store')->middleware(['auth']);
+    Route::delete('/admin/images/{id}/delete', 'destroy')->name('images.destroy')->middleware(['auth']);
 
 });
 
@@ -88,10 +88,10 @@ Route::controller(FileController::class)->group(function () {
     Route::get('/files/{id}/download', 'download')->name('files.download');
 
     // Admin File Routes
-    Route::get('/admin/files', 'index')->name('admin.files.index');
-    Route::get('/admin/files/create', 'create')->name('admin.files.create');
-    Route::post('/admin/files/create', 'store')->name('admin.files.store');
-    Route::delete('/admin/files/{id}/delete', 'destroy')->name('admin.files.destroy');
+    Route::get('/admin/files', 'index')->name('admin.files.index')->middleware(['auth']);
+    Route::get('/admin/files/create', 'create')->name('admin.files.create')->middleware(['auth']);
+    Route::post('/admin/files/create', 'store')->name('admin.files.store')->middleware(['auth']);
+    Route::delete('/admin/files/{id}/delete', 'destroy')->name('admin.files.destroy')->middleware(['auth']);
 
 });
 
@@ -104,8 +104,8 @@ Route::controller(CommentController::class)->group(function() {
     Route::post('/comments/create', 'store')->name('comments.store');
 
     // Admin Comment Routes
-    Route::get('/admin/comments', 'adminIndex')->name('admin.comments.index');
-    Route::delete('/admin/comments/{id}/delete', 'destroy')->name('admin.comments.destroy');
+    Route::get('/admin/comments', 'adminIndex')->name('admin.comments.index')->middleware(['auth']);
+    Route::delete('/admin/comments/{id}/delete', 'destroy')->name('admin.comments.destroy')->middleware(['auth']);
 
 });
